fix: avoid leaving zombie processes on unix
In `spawn_detached()`, the spawned process isn't waited on, which makes
zombie processes. This is fixed by calling `wait()` on the `Child`.

While this isn't done in the original function taken from the `open`
crate, it is done in `Alacritty`'s implementation, which is what the
author of `open` based the function on anyway.